requirements.txt
django>=5.0.0
anthropic>=0.39.0
openai>=1.0.0
mcp[cli]>=1.0.0
requests>=2.32.0
python-dotenv>=1.0.0
